BCDA-9633: add health_check var to ecs_service module (#354)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-9633

Related tickets:
[bcda-ops](https://github.com/CMSgov/bcda-ops/pull/1303)
[bcda-app](https://github.com/CMSgov/bcda-app/pull/1276)

## 🛠 Changes

<!-- What was added, updated, or removed in this PR? -->

- added an optional health check variable for the ecs service module

## ℹ️ Context

<!-- Why were these changes made? Add background context suitable for a
non-technical audience. -->

The BCDA worker service runs containers that don't have a target group
or a load balancer. It would be helpful to have a container health check
that can monitor unhealthy containers.

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

<!-- How were the changes verified? Did you fully test the acceptance
criteria in the ticket? Provide reproducible testing instructions and
screenshots if applicable. -->

Tested with bcda ecs services in dev environment.